Mantener usuario, eliminar plato, editar plato, agregar plato arreglados...

+++ src/app/component/orden/orden.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { OrdersService } from '../../service/orders/orders.service';
 import { PlatesService } from '../../service/plates/plates.service';
 import { ProductsService } from '../../service/products/products.service';
+import { ProductCustomizationComponent } from '../../component/product-customization/product-customization.component';
+
 import * as firebase from 'firebase';
 import { Router } from '@angular/router';
 
@@ -13,6 +15,10 @@ import { Router } from '@angular/router';
 export class OrdenComponent implements OnInit {
 
 	public orders = [];
+  public idOrder: any;
+  private eliminar: boolean = false;
+  private idPlatoEliminar: any;
+  private productCustomization: ProductCustomizationComponent;
 
   constructor(
   	private ordersService: OrdersService,
@@ -23,35 +29,63 @@ export class OrdenComponent implements OnInit {
 
 
   onShowProductCustomization(name: String) {
+    this.platesService.setAgregar(false);
+    this.platesService.setEditar(true);
     this.router.navigate([`/product-customization/${name}`]);
   }
 
   eliminarPlato(idPlate: string){
-  	this.ordersService.getOrders().subscribe((orderSnapshot) => {
-  		orderSnapshot.forEach((orderData: any) => {
-  			if(orderData.payload.doc.data().reference_user === firebase.auth().currentUser.email && orderData.payload.doc.data().actual) {
-        		
-        		const arrayPlates = orderData.payload.doc.data().plates_references;
-	        	for(let i = 0; i < arrayPlates.length; i++){
-	        		if(arrayPlates[i] === idPlate){
-	        			this.platesService.deletePlate(idPlate).then(res => {
-	        				alert("¡Plato eliminado exitosamente!");
-	        			})
-	        			this.ordersService.getOrder(orderData.payload.doc.id).update({
-	        				plates_references: firebase.firestore.FieldValue.arrayRemove(idPlate)
-	        			})
-	        		}
-	        	}
-        	}
-  		})
-  	})
+
+    var mensaje = confirm("¿Estás seguro de que deseas eliminar el producto? Si haces click en 'Aceptar', el mismo no podrá ser recuperado.");
+    if (mensaje) {
+
+      this.ordersService.getOrders().subscribe((orderSnapshot) => {
+        orderSnapshot.forEach((orderData: any) => {
+          if(orderData.payload.doc.data().reference_user === firebase.auth().currentUser.email && orderData.payload.doc.data().actual) {
+              
+              const arrayPlates = orderData.payload.doc.data().plates_references;
+              
+              for(let i = 0; i < arrayPlates.length; i++){
+                if(arrayPlates[i] === idPlate){
+                  this.platesService.deletePlate(idPlate).then(res => {
+                    alert("¡Plato eliminado exitosamente!");
+                  })
+                  this.ordersService.getOrder(orderData.payload.doc.id).update({
+                    plates_references: firebase.firestore.FieldValue.arrayRemove(idPlate)
+                  })
+                  break;
+                }
+              }
+          }
+        })
+      })
+
+
+      /*this.platesService.getPlates().subscribe((plateSnapshot) => {
+        plateSnapshot.forEach((plateData: any) => {
+          firebase.firestore().collection('/plates/').doc(plateData.payload.doc.id).onSnapshot((data) => {
+            if (data.id === idPlate) {
+              this.ordersService.getOrder(this.idOrder).update({
+                plates_references: firebase.firestore.FieldValue.arrayRemove(idPlate)
+              })
+              this.platesService.deletePlate(plateData.payload.doc.id);
+              alert("¡El plato ha sido eliminado con éxito de la orden!");
+            }
+          })
+        })
+      })*/
+    }
   }
   
   ngOnInit() {
   	this.ordersService.getOrders().subscribe((orderSnapshot) => {
+      this.orders = []
       orderSnapshot.forEach((orderData: any) => {
         if(orderData.payload.doc.data().reference_user === firebase.auth().currentUser.email && orderData.payload.doc.data().actual) {
+     
         	const arrayPlates = orderData.payload.doc.data().plates_references;
+          this.idOrder = orderData.payload.id
+
         	for(let i = 0; i < arrayPlates.length; i++){
 
         		this.platesService.getPlate(arrayPlates[i]).snapshotChanges().subscribe(dataPlate => {
